refactor: separate model modules into their corresponding blueprints
