+++ quotes_api/api/models/__init__.py
@@ -1,0 +1,5 @@
+"Quote api models initialization file."
+
+from quotes_api.api.models.quote import QuoteFields, Quote
+
+__all__ = ["QuoteFields", "Quote"]
