+++ quotes_api/api/models/quote.py
