+++ quotes_api/auth/models/__init__.py
@@ -1,0 +1,12 @@
+"""Auth models initialization file."""
+
+
+from quotes_api.auth.models.user import User, UserFields
+from quotes_api.auth.models.blacklist import TokenBlacklist, TokenBlacklistFields
+
+__all__ = [
+    "User",
+    "UserFields",
+    "TokenBlacklist",
+    "TokenBlacklistFields",
+]
