--- quotes_api/models/__init__.py
@@ -1,14 +1,0 @@
-"""Models initialization file."""
-
-from quotes_api.models.quote import Quote, QuoteFields
-from quotes_api.models.user import User, UserFields
-from quotes_api.models.blacklist import TokenBlacklist, TokenBlacklistFields
-
-__all__ = [
-    "Quote",
-    "QuoteFields",
-    "User",
-    "UserFields",
-    "TokenBlacklist",
-    "TokenBlacklistFields",
-]
